Add a firmware updater
Add a firmware updater for the Android recovery mode
(implemented as a shared library providing a custom script command 'firmware')
and a standalone command-line utility 'fwtool' excercising the various firmware
interfaces through the same code as the recovery updater.

The updater is currently doing only RW updates, even on not-Write-protected
machines, the RO firmware is not updated.

Change-Id: I01b617057812ce33fae3aacc83a48f33b7723bad
Signed-off-by: Vincent Palatin <vpalatin@chromium.org>