dragon: sensor_hub:  read sensor ring once.
Code was reading one event at a time. However, given the sensor
is opened as a blocking file, we would be waiting for the next event
in readEvents(), which is not allowed.

Given we are called by poll, call read only once to read as much event
as we can, and then process them.

Bug: 23297064

Change-Id: I6deb4532fff7ae5143e1e50b456e04b8e9a8a483
Signed-off-by: Gwendal Grignou <gwendal@google.com>