dragon: Bin the brightness to avoid excessive updates
Only provide the kernel with 16 brightness levels, instead of
the max_brightness discrete levels. This allows us to greatly
reduce the brightness updates that go on the wire, making the
brightness slider snappier.

It would be nice to only send each level to the kernel once
(instead of sending the same level a bunch of times). However,
we don't get notified when the panel goes on/off, so we can't
make a decision to resend a brightness when the panel comes up.
Fortunately the kernel can easily avoid persisting duplicate
brightnesses

Change-Id: Ib161cfeb8a30a55439ecdb3836ccaa4bacf200ed
Signed-off-by: Sean Paul <seanpaul@chromium.org>