Dump the charger chip state into bug reports
Read and record the I2C registers of the BQ25892 charger chip and
the EC charge manager / PD ports state, in order to debug
the input current limit issue.

As accessing the EC to retrieve this information requires priviledges and
dumpstate is run as the 'shell' user, send a property to the init.rc to trigger
the charge information dump into a file from there, then read this file
from dumpstate.

Bug: 24263374
Change-Id: I40dc4045f8cf1dbd44b8cbddcc89c42c058acede