dragon: create DragonKeyboard APK
Similarly to Volantis keyboard, Dragon (Pixel C) keyboard is lacking
several keys, such as ESC, left and right brackets, and backslash and
therefore requires custom key mapping. It is unclear if we will be using
this keyboard with anything but Dragon, so let's add it as
dragon-specific APK.

The mapping for US keyboard as follows:

AltR + 1 => ESC
[Shift +] AltR + 2 => ` and ~
[Shift +] AltR + O => [ and {
[Shift +] AltR + P => ] and }
[Shift +] AltR + = => \ and |

US international are mapped similarly with additional Crtl modifier,
since AltR modifier is needed for international symbols:

Ctrl + [Shift +] AltR + O => [ and {

Most international keyboards have [, ], {, and } characters on their
keys, the ones that do not have them are mapped similarly to US. Also
many international keyboards have < and > symbols mapped as AltR + ','
or AltR + '.'. ESC is mapped via AltR + 1 or AlrR + Ctrl + 1, depending
on whether AltR is reserved for an international symbol or not.

Bug: 24504154

Change-Id: I92477a0921c415f63a6edef30308c26a3c8c221d