dragon: sensor: Disable all sensors before closing the iio ring.
If we don't silence the sensors, the EC will keep posting events, so
we can't close the iio ring. It leads to a timeout in the kernel,
because we can not remove the buffer irq.

BUG=chrome-os-partner:44354
TEST=I can not reproduce the exact panic, but if I leave the sensors
activated, doing:
while [ true ]
do
  for i in 0 1
  do
      sleep 1
      echo $i > enable
   done
done
in /sys/bus/iio/devices/iio:device0/buffer crashes the system.
When all sensors are disabled, it does not happen.

Change-Id: Ie9d66df181055c5676328bc28ba9baa6fcd0bf18
Signed-off-by: Gwendal Grignou <gwendal@google.com>