dragon: sensor: Do not trust timestamp for activities event
If an activity triggered a resume, the activity event can be collected
before the RTC clock has been updated with the time:
PMIC is on i2c bus 5, while EC is on bus 1.
Therefore, the timestamp will not include the suspended delta,
failing CTS Significant Motion Test.

There is still an issue for sensors events that are collected while
in suspend: their timestamp will not include the delta as well
if they are collected before the RTC update.

BUG=chrome-os-partner:46724
TEST=With fix, test testAPWakeUpOnSMDTrigger pass.

Change-Id: I97eae3d83dd5f0519cbc048947214e2a521bb5fa
Signed-off-by: Gwendal Grignou <gwendal@google.com>