dragon: update Bluetooth firmware to BCM4350C0_003.001.012.0364.0754.hcd
Broadcom's comments:

Based on the OTA logs after repro'ing the issue locally with the co-ex
team, we observed that when the issue is happening, due to reduced BT
sensitivity, the BLE packets are being delayed in being received. This
was leading to longer BLE KEY_RELEASE packets and hence the key repeats.

In this test firmware, we have 2 patches :
1. Added patch to internally pull down BT_GPIO4, BT_GPIO5, BT_GPIO6 
   as we found it is affecting AFH.
2. Disabled ETSI for this firmware as A44 is a class 2 device and it is
   not needed.

TEST: paired the Bluetooth keyboard and verified that the RSSI doesn't
drop as much as before.
Change-Id: Idb7e5bcdeee44fa51449d3caf08a8659b67b60a0
Signed-off-by: Myles Watson <mylesgw@google.com>