dragon: recovery: updater: update_fw: Check active-firmware-version against RW bootloader version in OTA
All the bootloader images checked in for OTA have a special RO locked
down at 226 (RO used by factory). Thus, RO and RW versions do not
match. While performing an OTA check the active bootloader version on
device against the RW bootloader version in the OTA.

BUG=b:25690378
TEST=Verified using "fwtool update bootloader.img" that bootloader on
a device with active firmware version as 226 is updated to 251.

Change-Id: I9dae84e4018433a6a8aabbd1d3baa22a547cd8b5
Signed-off-by: Furquan Shaikh <furquan@google.com>