dragon: Add DragonKeyboardFirmwareUpdater service.
DragonKeyboardFirmwareUpdater is a built-in system service constantly running
in background and detecting if the connected Pixel C keyboard needs to be
updated. This service imports dfu library as a dependency to perform keyboard
firmware update. Except for the hacks mentioned in dfu/README.google, the
remaining codes are original.

This is the initial commit of DragonKeyboardFirmwareUpdater. The below
issues need to be resolved in the following CLs:
(1) android.permission.ACCESS_COARSE_LOCATION should be granted without user
interaction.
(2) The UI needs to be polished according to approved UI designed.
(3) Adding Pixel C Keyboard entry in System Settings to provide an entry for
postponed update.
(4) Remove hacks in dfu library.

Bug:25369339
Change-Id: I28b9b8ae671f2fb4a417fc45a9709aa6858f5c3e