vboot_interface: Make NVRAM update more robust
We are seeing a few issues in the field where NVRAM seems to be all
empty. After running some experiments it was observed that a race
condition could occur if device crashes/reboots in an unclean way
while fwtool is updating the NVRAM.

Currently, fwtool erases whole of NVRAM and writes back to it with a
new entry added. This results in a read-erase-write operation where
fwtool has no control if kernel panics. Thus, instead of erasing the
whole NVRAM every boot, use the following logic as used in firmware
for updating NVRAM:
1. Check if the entry can be overwritten on the last-used entry
(i.e. only 1->0 transitions).
2. If yes, update only the last used entry.
3. If not, check if there is space for a new entry in NVRAM.
4. If yes, write to the entry after the last-used entry.
5. If not, erase NVRAM and update values in the first entry.

With this, the NVRAM does an erase-write cycle only one in 4k
reboots. During the rest of reboot cycles, we only perform a write
operation. Though this does not completely eliminate the race
condition, it makes the process more robust. Also, since we sync both
slots A and B after a successful OTA and boot to OS, falling back to
other slot should not harm as well.

BUG=b/26193350
TEST=Verified that NVRAM is updated as per the logic mentioned
above. Erase of NVRAM section is done only after all the entries are consumed.

Change-Id: I00493614f349da0336c110afaca53d22736b2834
Signed-off-by: Furquan Shaikh <furquan@google.com>