allow mediacodec to read /system/lib dir
Very strange that this process needs to read the directory. It appears
to be listing the contents of the directory in order to find a library
name that contains "egl." Instead of the more typical method of opening
a known file name - in this case libegl.so.

Addresses:
avc: denied { read } for name="egl" dev="dm-1" ino=131
scontext=u:r:mediacodec:s0 tcontext=u:object_r:system_file:s0 tclass=dir

Bug: 27530325
Change-Id: I3f6bc4a06eec6f8ef1e08f5f1897c8c03a035075