Disable building of TLK, to unbreak master build.
TLK is currently pulling in bionic headers, even though it doesn't use
bionic. As a result, changes in those headers in AOSP have broken the
build. TLK should not depend on bionic at all, and should have
everything needed to build in its own headers. Fixing that is
non-trivial, so this disables building of TLK while the ryu team and
nvidia fix it properly.

Change-Id: I770609a445d184f5e0f60bbf1963e994b5bba6ce